refactor: source files into src dir
