+++ src/graph.cpp
@@ -4,9 +4,9 @@
 #include <chrono>
 #include <sys/mman.h>
 
-#include "include/graph.h"
-#include "include/util.h"
-#include "include/graph_worker.h"
+#include "../include/graph.h"
+#include "../include/util.h"
+#include "../include/graph_worker.h"
 
 Graph::Graph(uint64_t num_nodes): num_nodes(num_nodes) {
 #ifdef VERIFY_SAMPLES_F
