+++ src/l0_sampling/sketch.cpp
@@ -1,4 +1,4 @@
-#include "../include/l0_sampling/sketch.h"
+#include "../../include/l0_sampling/sketch.h"
 #include <cassert>
 #include <iostream>
