+++ src/util.cpp
@@ -1,8 +1,8 @@
 #include <stdexcept>
 #include <boost/multiprecision/cpp_int.hpp>
-#include "include/util.h"
-#include "include/graph_worker.h"
-#include "include/graph.h"
+#include "../include/util.h"
+#include "../include/graph_worker.h"
+#include "../include/graph.h"
 
 const char *config_file = "streaming.conf";
 using uint128_t = boost::multiprecision::uint128_t;
